requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
opencv-python==4.8.1.78
numpy>=1.26.0
scikit-learn>=1.3.2
face-recognition>=1.3.5
librosa>=0.10.0
pyttsx3==2.90
pydub==0.25.1
python-decouple==3.8
mysql-connector-python==8.2.0
openpyxl==3.11.0
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1